refactor: move all flag arguments manage logic to pkg /gflag
