+++ app_test.go
@@ -22,7 +22,7 @@ var (
 		Desc: "an simple command",
 		Func: func(c *gcli.Command, args []string) error {
 			dump.Println(c.Path(), args)
-			c.SetValue("simple", "simple command")
+			c.Set("simple", "simple command")
 			return nil
 		},
 	}
@@ -45,7 +45,7 @@ var (
 					Name: "sub1",
 					Desc: "desc for top1.sub1",
 					Func: func(c *gcli.Command, args []string) error {
-						c.SetValue("msg", c.App().GetVal("top1:sub1"))
+						c.Set("msg", c.App().Get("top1:sub1"))
 						return nil
 					},
 				},
@@ -291,12 +291,12 @@ func TestApp_Run_subcommand(t *testing.T) {
 	is := assert.New(t)
 	id := "top1:sub1"
 
-	appWithMl.SetValue(id, "TestApp_Run_subcommand")
+	appWithMl.Set(id, "TestApp_Run_subcommand")
 	appWithMl.Run([]string{"top1", "sub1"})
 
 	c := appWithMl.FindCommand(id)
 	is.NotEmpty(c)
-	is.Eq("TestApp_Run_subcommand", c.GetVal("msg"))
+	is.Eq("TestApp_Run_subcommand", c.Get("msg"))
 }
 
 func TestApp_Run_by_cmd_ID(t *testing.T) {
@@ -452,6 +452,6 @@ func (uc *UserCommand) Config(c *gcli.Command) {
 	c.StrOpt(&uc.opt1, "opt", "o", "", "desc")
 }
 
-func (uc *UserCommand) Execute(c *gcli.Command, args []string) error {
+func (uc *UserCommand) Execute(_ *gcli.Command, _ []string) error {
 	return nil
 }
